+++ paper/src/main/java/dev/qixils/crowdcontrol/plugin/paper/commands/EnchantmentCommand.java
@@ -7,10 +7,16 @@ import dev.qixils.crowdcontrol.socket.Response;
 import lombok.Getter;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.PlayerInventory;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Comparator;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 
 @Getter
 public class EnchantmentCommand extends ImmediateCommand {
@@ -32,17 +38,39 @@ public class EnchantmentCommand extends ImmediateCommand {
 				.message("No items could be enchanted");
 		for (Player player : players) {
 			int level = enchantment.getMaxLevel();
-			ItemStack item = player.getInventory().getItemInMainHand();
-			if (item.getType().isEmpty()) {
-				item = player.getInventory().getItemInOffHand();
-				if (item.getType().isEmpty())
-					continue;
+			PlayerInventory inv = player.getInventory();
+			// get the equipped item that supports this enchantment and has the lowest level of it
+			Map<EquipmentSlot, Integer> levelMap = new HashMap<>(EquipmentSlot.values().length);
+			for (EquipmentSlot slot : EquipmentSlot.values()) {
+				ItemStack item = inv.getItem(slot);
+				// ensure this item:
+				// A) isn't null
+				// B) is not empty
+				// C) supports the requested enchantment
+				// D) would actually benefit from upgrading this enchantment if applicable
+				//    (i.e. this prevents Silk Touch from being "upgraded" to level 2)
+				if (item != null
+						&& !item.getType().isEmpty()
+						&& enchantment.canEnchantItem(item)
+						&& (
+						enchantment.getMaxLevel() != enchantment.getStartLevel()
+								|| item.getEnchantmentLevel(enchantment) != enchantment.getMaxLevel()
+				)
+				) {
+					levelMap.put(slot, item.getEnchantmentLevel(enchantment));
+				}
 			}
+			EquipmentSlot slot = levelMap.entrySet().stream()
+					.min(Comparator.comparingInt(Entry::getValue))
+					.map(Entry::getKey).orElse(null);
+			if (slot == null)
+				continue;
+			// add enchant
+			ItemStack item = inv.getItem(slot);
+			assert item != null : "Selected item has disappeared from the player's inventory";
 			int curLevel = item.getEnchantmentLevel(enchantment);
 			if (curLevel >= level)
 				level = curLevel + 1;
-			else if (curLevel == 0 && !enchantment.getItemTarget().includes(item))
-				continue;
 			item.addUnsafeEnchantment(enchantment, level);
 			result.type(Response.ResultType.SUCCESS);
 		}