+++ sponge-7/src/main/java/dev/qixils/crowdcontrol/plugin/sponge7/commands/EnchantmentCommand.java
@@ -8,16 +8,21 @@ import dev.qixils.crowdcontrol.socket.Response;
 import dev.qixils.crowdcontrol.socket.Response.ResultType;
 import lombok.Getter;
 import org.jetbrains.annotations.NotNull;
-import org.spongepowered.api.data.manipulator.mutable.item.EnchantmentData;
+import org.spongepowered.api.data.key.Keys;
 import org.spongepowered.api.data.type.HandTypes;
-import org.spongepowered.api.data.value.mutable.ListValue;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.item.enchantment.Enchantment;
 import org.spongepowered.api.item.enchantment.EnchantmentType;
 import org.spongepowered.api.item.inventory.ItemStack;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Optional;
 
 @Getter
@@ -35,6 +40,12 @@ public class EnchantmentCommand extends ImmediateCommand {
 		this.displayName = "Apply " + enchantmentType.getTranslation().get();
 	}
 
+	private int getCurrentLevel(ItemStack item) {
+		return item.get(Keys.ITEM_ENCHANTMENTS)
+				.flatMap(enchs -> enchs.stream().filter(ench -> ench.getType().equals(enchantmentType)).findFirst())
+				.map(Enchantment::getLevel).orElse(0);
+	}
+
 	@NotNull
 	@Override
 	public Response.Builder executeImmediately(@NotNull List<@NotNull Player> players, @NotNull Request request) {
@@ -42,26 +53,38 @@ public class EnchantmentCommand extends ImmediateCommand {
 				.type(ResultType.RETRY)
 				.message("No items could be enchanted");
 		for (Player player : players) {
-			Optional<ItemStack> optionalStack = player.getItemInHand(HandTypes.MAIN_HAND);
-			if (!optionalStack.isPresent())
-				optionalStack = player.getItemInHand(HandTypes.OFF_HAND);
+			// get the equipped item that supports this enchantment and has the lowest level of it
+			Map<Slot, Integer> levelMap = new HashMap<>(Slot.values().length);
+			for (Slot slot : Slot.values()) {
+				Optional<ItemStack> optionalStack = slot.getItem(player);
+				if (!optionalStack.isPresent())
+					continue;
+				ItemStack item = optionalStack.get();
+				if (item.isEmpty())
+					continue;
+				if (!enchantmentType.canBeAppliedToStack(item))
+					continue;
+				int curLevel = getCurrentLevel(item);
+				if (enchantmentType.getMaximumLevel() == enchantmentType.getMinimumLevel() && curLevel == enchantmentType.getMaximumLevel())
+					continue;
+				levelMap.put(slot, curLevel);
+			}
+			Slot slot = levelMap.entrySet().stream()
+					.min(Comparator.comparingInt(Entry::getValue))
+					.map(Entry::getKey).orElse(null);
+			if (slot == null)
+				continue;
+			Optional<ItemStack> optionalStack = slot.getItem(player);
 			if (!optionalStack.isPresent())
 				continue;
 
+			// misc instantiation
 			ItemStack item = optionalStack.get();
-			if (!item.getOrCreate(EnchantmentData.class).isPresent())
-				continue;
-
-			Optional<EnchantmentData> optionalEnchantmentData = item.getOrCreate(EnchantmentData.class);
-			if (!optionalEnchantmentData.isPresent())
-				continue;
-
-			Enchantment toAdd = Enchantment.of(enchantmentType, maxLevel);
-			EnchantmentData enchantmentData = optionalEnchantmentData.get();
-			ListValue<Enchantment> enchantments = enchantmentData.enchantments();
+			List<Enchantment> enchantments = new ArrayList<>(item.get(Keys.ITEM_ENCHANTMENTS).orElseGet(Collections::emptyList));
 			Iterator<Enchantment> iterator = enchantments.iterator();
-			boolean canAdd = enchantmentType.canBeAppliedToStack(item);
 
+			// determine enchantment level
+			Enchantment toAdd = Enchantment.of(enchantmentType, maxLevel);
 			while (iterator.hasNext()) {
 				Enchantment enchantment = iterator.next();
 				if (!enchantment.getType().equals(enchantmentType))
@@ -70,19 +93,86 @@ public class EnchantmentCommand extends ImmediateCommand {
 				if (curLevel >= maxLevel)
 					toAdd = Enchantment.of(enchantmentType, curLevel + 1);
 				iterator.remove();
-				canAdd = true;
-				break;
 			}
 
-			if (!canAdd)
-				continue;
-
+			// add enchant
 			enchantments.add(toAdd);
-			// I think these are necessary
-			enchantmentData.set(enchantments);
-			item.offer(enchantmentData);
+			item.offer(Keys.ITEM_ENCHANTMENTS, enchantments);
 			response.type(ResultType.SUCCESS).message("SUCCESS");
 		}
 		return response;
+	}
+
+	private enum Slot {
+		MAIN_HAND {
+			@Override
+			public Optional<ItemStack> getItem(Player player) {
+				return player.getItemInHand(HandTypes.MAIN_HAND);
+			}
+
+			@Override
+			public void setItem(Player player, ItemStack item) {
+				player.setItemInHand(HandTypes.MAIN_HAND, item);
+			}
+		},
+		OFF_HAND {
+			@Override
+			public Optional<ItemStack> getItem(Player player) {
+				return player.getItemInHand(HandTypes.OFF_HAND);
+			}
+
+			@Override
+			public void setItem(Player player, ItemStack item) {
+				player.setItemInHand(HandTypes.OFF_HAND, item);
+			}
+		},
+		HELMET {
+			@Override
+			public Optional<ItemStack> getItem(Player player) {
+				return player.getHelmet();
+			}
+
+			@Override
+			public void setItem(Player player, ItemStack item) {
+				player.setHelmet(item);
+			}
+		},
+		CHESTPLATE {
+			@Override
+			public Optional<ItemStack> getItem(Player player) {
+				return player.getChestplate();
+			}
+
+			@Override
+			public void setItem(Player player, ItemStack item) {
+				player.setChestplate(item);
+			}
+		},
+		LEGGINGS {
+			@Override
+			public Optional<ItemStack> getItem(Player player) {
+				return player.getLeggings();
+			}
+
+			@Override
+			public void setItem(Player player, ItemStack item) {
+				player.setLeggings(item);
+			}
+		},
+		BOOTS {
+			@Override
+			public Optional<ItemStack> getItem(Player player) {
+				return player.getBoots();
+			}
+
+			@Override
+			public void setItem(Player player, ItemStack item) {
+				player.setBoots(item);
+			}
+		};
+
+		public abstract Optional<ItemStack> getItem(Player player);
+
+		public abstract void setItem(Player player, ItemStack item);
 	}
 }