+++ sponge-7/src/main/java/dev/qixils/crowdcontrol/plugin/sponge7/commands/GiveItemCommand.java
@@ -6,8 +6,10 @@ import dev.qixils.crowdcontrol.plugin.sponge7.utils.Sponge7TextUtil;
 import dev.qixils.crowdcontrol.socket.Request;
 import dev.qixils.crowdcontrol.socket.Response;
 import lombok.Getter;
+import org.jetbrains.annotations.Blocking;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.api.data.key.Keys;
+import org.spongepowered.api.entity.Entity;
 import org.spongepowered.api.entity.EntityTypes;
 import org.spongepowered.api.entity.Item;
 import org.spongepowered.api.entity.living.player.Player;
@@ -33,22 +35,31 @@ public class GiveItemCommand extends ImmediateCommand {
 		this.displayName = "Give " + item.getTranslation().get();
 	}
 
+	@Blocking
+	public static void giveItemTo(SpongeCrowdControlPlugin plugin, Entity player, ItemStack itemStack) {
+		Item entity = (Item) player.getLocation().createEntity(EntityTypes.ITEM);
+		entity.offer(Keys.REPRESENTED_ITEM, itemStack.createSnapshot());
+		// seems like Sponge 7 doesn't support any of the other keys used in the Paper impl
+		// so this is a note to add them back in the Sponge 8 impl [API8]
+
+		// give entity a cause & spawn it
+		try (StackFrame frame = plugin.getGame().getCauseStackManager().pushCauseFrame()) {
+			frame.addContext(EventContextKeys.SPAWN_TYPE, SpawnTypes.PLUGIN);
+			player.getWorld().spawnEntity(entity);
+		}
+	}
+
+	private void giveItemTo(Entity player, ItemStack itemStack) {
+		giveItemTo(plugin, player, itemStack);
+	}
+
 	@NotNull
 	@Override
 	public Response.Builder executeImmediately(@NotNull List<@NotNull Player> players, @NotNull Request request) {
+		ItemStack itemStack = ItemStack.of(item);
 		sync(() -> {
 			for (Player player : players) {
-				Item entity = (Item) player.getLocation().createEntity(EntityTypes.ITEM);
-				entity.offer(Keys.REPRESENTED_ITEM, ItemStack.of(item).createSnapshot());
-				// seems like Sponge 7 doesn't support any of the other keys used in the Paper impl
-				// so this is a note to add them back in the Sponge 8 impl [API8]
-
-				// give entity a cause & spawn it
-				try (StackFrame frame = plugin.getGame().getCauseStackManager().pushCauseFrame()) {
-					frame.addContext(EventContextKeys.SPAWN_TYPE, SpawnTypes.PLUGIN);
-					player.getWorld().spawnEntity(entity);
-				}
-
+				giveItemTo(player, itemStack);
 				// workaround to limit the circulation of end portal frames in the economy
 				if (item.equals(ItemTypes.END_PORTAL_FRAME))
 					break;